+++ src/service/http_handlers.go
@@ -418,7 +418,8 @@ GET /export
 returns: JSON Export of current state
 
 This endpoint returns a JSON snapshot of the state, containing all the accounts,
-including smart-contracts with their storage.
+including smart-contracts with their storage. The result can be reused as a
+genesis file.
 */
 func exportHandler(w http.ResponseWriter, r *http.Request, m *Service) {
 	m.logger.Debug("GET export")